requirements.txt
discord.py
bs4
python-dotenv
requests
fuzzywuzzy
imgkit
Pillow
